# backend/docs/swagger.yaml
basePath: /
definitions:
  models.Meeting:
    properties:
      allMembers:
        type: boolean
      createdAt:
        type: string
      createdBy:
        type: string
      date:
        type: string
      description:
        type: string
      duration:
        type: integer
      id:
        type: string
      participants:
        items:
          type: string
        type: array
      time:
        type: string
      title:
        type: string
    type: object
  models.MeetingResponse:
    properties:
      allMembers:
        type: boolean
      createdAt:
        type: string
      createdBy:
        $ref: '#/definitions/models.User'
      date:
        type: string
      description:
        type: string
      duration:
        type: integer
      id:
        type: string
      participants:
        items:
          $ref: '#/definitions/models.User'
        type: array
      time:
        type: string
      title:
        type: string
    type: object
  models.User:
    properties:
      bio:
        type: string
      email:
        type: string
      id:
        type: string
      lastActive:
        type: string
      nama:
        type: string
      password:
        type: string
      profileImage:
        type: string
      role:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for CoEmotion meeting management system
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: CoEmotion API
  version: "1.0"
paths:
...
  /api/meetings:
    get:
      description: Get list of all meetings
      produces:
      - application/json
      responses:
        "200":
          description: List of meetings
          schema:
            items:
              $ref: '#/definitions/models.Meeting'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get all meetings
      tags:
      - Meetings
    post:
      consumes:
      - application/json
      description: Create a new meeting with the provided details
      parameters:
      - description: Meeting data
        in: body
        name: meeting
        required: true
        schema:
          $ref: '#/definitions/models.Meeting'
      produces:
      - application/json
      responses:
        "201":
          description: Meeting created successfully
          schema:
            $ref: '#/definitions/models.Meeting'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create a new meeting
      tags:
      - Meetings
  /api/meetings/{id}:
    delete:
      description: Delete a meeting by meeting ID
      parameters:
      - description: Meeting ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Meeting deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Meeting not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete meeting
      tags:
      - Meetings
    get:
      description: Get meeting information by meeting ID
      parameters:
      - description: Meeting ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Meeting information
          schema:
            $ref: '#/definitions/models.MeetingResponse'
        "404":
          description: Meeting not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get meeting by ID
      tags:
      - Meetings
    put:
      consumes:
      - application/json
      description: Update meeting information by meeting ID
      parameters:
      - description: Meeting ID
        in: path
        name: id
        required: true
        type: string
      - description: Meeting update data
        in: body
        name: meeting
        required: true
        schema:
          $ref: '#/definitions/models.Meeting'
      produces:
      - application/json
      responses:
        "200":
          description: Meeting updated successfully
          schema:
            $ref: '#/definitions/models.Meeting'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Meeting not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update meeting
      tags:
      - Meetings
  /api/meetings/today:
    get:
      description: Get list of meetings scheduled for today
      produces:
      - application/json
      responses:
        "200":
          description: List of today's meetings
          schema:
            items:
              $ref: '#/definitions/models.Meeting'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get today's meetings
      tags:
      - Meetings
  /api/meetings/upcoming:
    get:
      description: Get list of meetings that are scheduled for today and haven't started
        yet, or are in the future
      produces:
      - application/json
      responses:
        "200":
          description: List of upcoming meetings
          schema:
            items:
              $ref: '#/definitions/models.Meeting'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get upcoming meetings
      tags:
      - Meetings
...
schemes:
- http
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"